Finalize cleanup based on review
