+++ frontend/src/modules/Intersection/DataProviderFramework/utils/extensionLengthUtils.ts
@@ -1,9 +1,12 @@
 import { IntersectionType } from "@framework/types/intersection";
 import type { IntersectionSettingValue } from "@modules/_shared/DataProviderFramework/settings/implementations/IntersectionSetting";
 
-export function createValidExtensionLength(intersection: IntersectionSettingValue | null): number {
+export function createValidExtensionLength(
+    intersection: IntersectionSettingValue | null,
+    fallbackExtensionLength = 0,
+): number {
     if (intersection?.type === IntersectionType.WELLBORE) {
-        return intersection.extensionLength ?? 0;
+        return intersection.extensionLength ?? fallbackExtensionLength;
     }
-    return 0;
+    return fallbackExtensionLength;
 }
