+++ frontend/src/modules/Intersection/view/atoms/baseAtoms.ts
@@ -2,7 +2,7 @@ import { atom } from "jotai";
 
 import type { ViewStateMap } from "@modules/Intersection/typesAndEnums";
 
-import type { ViewLink } from "../components/ViewLinkManager";
+import type { ViewLink } from "../typesAndEnums";
 
 export const viewLinksAtom = atom<ViewLink[] | null>(null);
 export const viewStateMapAtom = atom<ViewStateMap | null>(null);
