+++ frontend/src/modules/_shared/DataProviderFramework/settings/SettingRegistry/_registerAllSettings.ts
@@ -105,7 +105,7 @@ SettingRegistry.registerSetting(Setting.GRID_LAYER_K, "Grid Layer K", NumberRang
 SettingRegistry.registerSetting(Setting.GRID_LAYER_RANGE, "Grid Ranges", GridLayerRangeSetting);
 SettingRegistry.registerSetting(Setting.GRID_NAME, "Grid Name", DropdownStringSetting);
 SettingRegistry.registerSetting(Setting.INTERSECTION, "Intersection", IntersectionSetting, {
-    customConstructorParameters: [{ extensionLength: { min: 0, max: 5000, defaultValue: 500 } }],
+    customConstructorParameters: [{ extensionLengthConfig: { min: 0, max: 5000, defaultValue: 500 } }],
 });
 SettingRegistry.registerSetting(Setting.OPACITY_PERCENT, "Color Opacity [%]", SliderNumberSetting, {
     customConstructorParameters: [{ minMax: { min: 0, max: 100 }, step: 1 }],
