+++ frontend/src/modules/_shared/components/EsvIntersection/utilityComponents/Toolbar.tsx
@@ -40,8 +40,8 @@ export type ViewLinkOption = {
 };
 
 export type ToolbarProps = {
-    viewLinks?: ViewLinkOption[];
-    unlinkedViews?: ViewOption[];
+    viewLinks?: readonly ViewLinkOption[];
+    unlinkedViews?: readonly ViewOption[];
     visible: boolean;
     zFactor: number;
     gridVisible: boolean;
